migrate test3 and test4 to gtest
