+++ deque/CMakeLists.txt
@@ -7,5 +7,6 @@ find_package(GTest MODULE REQUIRED)
 
 include_directories(${GTEST_INCLUDE_DIRS})
 
-add_executable(DequeTest test_main.cpp test_1.cpp mocks/Integer.h mocks/Integer.cpp test_2.cpp mocks/Matrix.hpp mocks/Bint.cpp mocks/Bint.h mocks/DynamicType.cpp mocks/DynamicType.h)
+add_executable(DequeTest test_main.cpp basic_test.cpp mocks/Integer.h mocks/Integer.cpp random_test.cpp mocks/Matrix.hpp mocks/Bint.cpp mocks/Bint.h mocks/DynamicType.cpp mocks/DynamicType.h series_test.cpp mocks/Int.hpp)
+add_executable(DequeBench bench_main.cpp)
 target_link_libraries(DequeTest PRIVATE GTest::GTest GTest::Main)
