--- deque/data/four.memcheck/answer.txt
@@ -1,8 +1,0 @@
-test start:
-test1: push & pop                    Accept
-test2: at & [] & front & back        Accept
-test3: itetator operation            Accept
-test4: const_itetator operation      Accept
-test5: erase & insert                Accept
-test6: clear & copy & assignment     Accept
-test7: complexity                    Accept
